- Fix errors with shipyard trade stations. - Added new API methods. - Added new admin commands:
Return a list of prefabs loaded in the save with information about the grid, can take some time to load this info, so this command run in background.
/definitions prefabs

Spawn a prefab by name at 250m in front of the player. (There are no validations about the colisions, so take care what you spawn)
/spawnprefab <PREFAB NAME>

- Add support to prefab in "OKI NPC Expansion Mod" to be sell in trade stations.
https://steamcommunity.com/sharedfiles/filedetails/?id=420119159